fix(connector): correct EGFX docs and coverage
Point EGFX users to the supported GraphicsPipelineClient registration path and clarify that disabling the flag only leaves EGFX unadvertised.

Move capability-bit coverage into the enabled core testsuite and decode the emitted Connect Initial PDU so normal CI exercises the assertion.
