+++ crates/ironrdp-connector/src/connection.rs
@@ -2022,95 +2022,4 @@ mod tests {
                 .contains(gcc::ClientEarlyCapabilityFlags::SUPPORT_MONITOR_LAYOUT_PDU)
         );
     }
-
-    fn dyn_vc_gfx_config(support_dyn_vc_gfx_protocol: bool) -> Config {
-        Config {
-            desktop_size: DesktopSize {
-                width: 1024,
-                height: 768,
-            },
-            monitor_layout: None,
-            desktop_scale_factor: 0,
-            enable_tls: true,
-            enable_credssp: false,
-            enable_standard_rdp_security: false,
-            credentials: Credentials::UsernamePassword {
-                username: "test".into(),
-                password: "test".into(),
-            },
-            domain: None,
-            client_build: 0,
-            client_name: "test".into(),
-            keyboard_type: gcc::KeyboardType::IBM_ENHANCED,
-            keyboard_subtype: 0,
-            keyboard_functional_keys_count: 12,
-            keyboard_layout: 0,
-            connection_type: gcc::ConnectionType::Lan,
-            ime_file_name: String::new(),
-            bitmap: None,
-            dig_product_id: String::new(),
-            client_dir: String::new(),
-            alternate_shell: String::new(),
-            work_dir: String::new(),
-            remote_application_mode: false,
-            rail_support_level: RailSupportLevel::empty(),
-            platform: MajorPlatformType::UNIX,
-            hardware_id: None,
-            request_data: None,
-            autologon: false,
-            enable_audio_playback: false,
-            enable_audio_capture: false,
-            performance_flags: Default::default(),
-            license_cache: None,
-            timezone_info: Default::default(),
-            compression_type: None,
-            enable_server_pointer: false,
-            pointer_software_rendering: false,
-            multitransport_flags: None,
-            support_dyn_vc_gfx_protocol,
-        }
-    }
-
-    fn early_capability_flags(config: &Config) -> gcc::ClientEarlyCapabilityFlags {
-        create_gcc_blocks(config, nego::SecurityProtocol::SSL, true, core::iter::empty())
-            .expect("GCC blocks")
-            .core
-            .optional_data
-            .early_capability_flags
-            .expect("early capability flags are always emitted")
-    }
-
-    /// The default must leave the wire bit clear: a client that has not wired up
-    /// an EGFX-capable DVC processor and gets this flag set anyway sees modern
-    /// Windows servers stop sending legacy bitmap updates, and the desktop goes
-    /// blank.
-    #[test]
-    fn dyn_vc_gfx_protocol_bit_is_absent_by_default() {
-        assert!(
-            !early_capability_flags(&dyn_vc_gfx_config(false))
-                .contains(gcc::ClientEarlyCapabilityFlags::SUPPORT_DYN_VC_GFX_PROTOCOL)
-        );
-    }
-
-    /// Opting in sets the bit the server negotiates the Graphics Pipeline on.
-    #[test]
-    fn dyn_vc_gfx_protocol_bit_is_set_when_enabled() {
-        assert!(
-            early_capability_flags(&dyn_vc_gfx_config(true))
-                .contains(gcc::ClientEarlyCapabilityFlags::SUPPORT_DYN_VC_GFX_PROTOCOL)
-        );
-    }
-
-    /// The opt-in is additive: it must not disturb the flags the connector
-    /// always advertises, nor the ones derived from other config.
-    #[test]
-    fn dyn_vc_gfx_protocol_opt_in_changes_exactly_one_bit() {
-        let without = early_capability_flags(&dyn_vc_gfx_config(false));
-        let with = early_capability_flags(&dyn_vc_gfx_config(true));
-
-        assert_eq!(
-            with.symmetric_difference(without),
-            gcc::ClientEarlyCapabilityFlags::SUPPORT_DYN_VC_GFX_PROTOCOL,
-        );
-    }
 }
