+++ crates/ironrdp-testsuite-core/tests/connector/autodetect.rs
@@ -10,12 +10,10 @@
 use std::borrow::Cow;
 
 use ironrdp_connector::MonotonicInstant;
-use ironrdp_connector::{ClientConnector, ClientConnectorState, Credentials, DesktopSize, Sequence as _, Written};
+use ironrdp_connector::{ClientConnector, ClientConnectorState, Sequence as _, Written};
 use ironrdp_core::{WriteBuf, encode_vec};
-use ironrdp_pdu::gcc;
 use ironrdp_pdu::mcs::{McsMessage, SendDataIndication};
 use ironrdp_pdu::rdp::autodetect::{AutoDetectReqPdu, AutoDetectRequest, AutoDetectResponse, AutoDetectRspPdu};
-use ironrdp_pdu::rdp::capability_sets::MajorPlatformType;
 use ironrdp_pdu::rdp::headers::{BasicSecurityHeader, BasicSecurityHeaderFlags};
 use ironrdp_pdu::rdp::server_license::{
     LicenseErrorCode, LicenseHeader, LicensePdu, LicensingErrorMessage, LicensingStateTransition, PreambleFlags,
@@ -27,58 +25,10 @@ const USER_CHANNEL_ID: u16 = 1002;
 const IO_CHANNEL_ID: u16 = 1003;
 const MESSAGE_CHANNEL_ID: u16 = 1004;
 
-fn test_config() -> ironrdp_connector::Config {
-    ironrdp_connector::Config {
-        desktop_size: DesktopSize {
-            width: 1024,
-            height: 768,
-        },
-        monitor_layout: None,
-        desktop_scale_factor: 0,
-        enable_tls: true,
-        enable_credssp: false,
-        enable_standard_rdp_security: false,
-        credentials: Credentials::UsernamePassword {
-            username: "test".into(),
-            password: "test".into(),
-        },
-        domain: None,
-        client_build: 0,
-        client_name: "test".into(),
-        keyboard_type: gcc::KeyboardType::IBM_ENHANCED,
-        keyboard_subtype: 0,
-        keyboard_layout: 0,
-        keyboard_functional_keys_count: 12,
-        connection_type: gcc::ConnectionType::Lan,
-        ime_file_name: String::new(),
-        bitmap: None,
-        dig_product_id: String::new(),
-        client_dir: String::new(),
-        platform: MajorPlatformType::UNIX,
-        hardware_id: None,
-        request_data: None,
-        autologon: false,
-        enable_audio_playback: false,
-        enable_audio_capture: false,
-        license_cache: None,
-        compression_type: None,
-        enable_server_pointer: false,
-        pointer_software_rendering: false,
-        multitransport_flags: None,
-        support_dyn_vc_gfx_protocol: false,
-        performance_flags: Default::default(),
-        timezone_info: Default::default(),
-        alternate_shell: String::new(),
-        work_dir: String::new(),
-        remote_application_mode: false,
-        rail_support_level: ironrdp_pdu::rdp::capability_sets::RailSupportLevel::SUPPORTED,
-    }
-}
-
 /// A client connector parked in `ConnectTimeAutoDetection` with a negotiated
 /// message channel, ready to receive the first PDU of that phase.
 fn connect_time_autodetect_connector() -> ClientConnector {
-    let mut connector = ClientConnector::new(test_config(), "127.0.0.1:12345".parse().unwrap());
+    let mut connector = ClientConnector::new(super::test_config(), "127.0.0.1:12345".parse().unwrap());
     connector.state = ClientConnectorState::ConnectTimeAutoDetection {
         io_channel_id: IO_CHANNEL_ID,
         user_channel_id: USER_CHANNEL_ID,
