+++ crates/ironrdp-testsuite-core/tests/connector/mod.rs
@@ -1,1 +1,54 @@
+use ironrdp_connector::{Credentials, DesktopSize};
+use ironrdp_pdu::gcc;
+use ironrdp_pdu::rdp::capability_sets::{MajorPlatformType, RailSupportLevel};
+
 mod autodetect;
+mod early_capabilities;
+
+fn test_config() -> ironrdp_connector::Config {
+    ironrdp_connector::Config {
+        desktop_size: DesktopSize {
+            width: 1024,
+            height: 768,
+        },
+        monitor_layout: None,
+        desktop_scale_factor: 0,
+        enable_tls: true,
+        enable_credssp: false,
+        enable_standard_rdp_security: false,
+        credentials: Credentials::UsernamePassword {
+            username: "test".into(),
+            password: "test".into(),
+        },
+        domain: None,
+        client_build: 0,
+        client_name: "test".into(),
+        keyboard_type: gcc::KeyboardType::IBM_ENHANCED,
+        keyboard_subtype: 0,
+        keyboard_layout: 0,
+        keyboard_functional_keys_count: 12,
+        connection_type: gcc::ConnectionType::Lan,
+        ime_file_name: String::new(),
+        bitmap: None,
+        dig_product_id: String::new(),
+        client_dir: String::new(),
+        platform: MajorPlatformType::UNIX,
+        hardware_id: None,
+        request_data: None,
+        autologon: false,
+        enable_audio_playback: false,
+        enable_audio_capture: false,
+        license_cache: None,
+        compression_type: None,
+        enable_server_pointer: false,
+        pointer_software_rendering: false,
+        multitransport_flags: None,
+        support_dyn_vc_gfx_protocol: false,
+        performance_flags: Default::default(),
+        timezone_info: Default::default(),
+        alternate_shell: String::new(),
+        work_dir: String::new(),
+        remote_application_mode: false,
+        rail_support_level: RailSupportLevel::SUPPORTED,
+    }
+}
